fix: Code review round 3 — ~80 issues across 8 phases (safety, logic, resources, perf, SDK, CI)
Comprehensive code review fixes organized in 8 phases:

Phase 1: NaN-safe sorting, overflow guards, naming/doc fixes, dead code removal
Phase 2: Camera intrinsics guards, 180° rotation singularity, NaN divergence checks,
         RANSAC seed from data, monotonic timestamp validation, Sampson Inf guard
Phase 3: BA convergence flag, rotation divisor, IDF dirty on remove, negative speed
         guard, Anchor event ordering, serialize deep copy, tracker resolution fix
Phase 4: WorkerBridge 6DoF tracker, terminate() race fix, init promise rejection,
         .free() cleanup, GPU resource dispose, HitTester destroy()
Phase 5: VecDeque for FIFO collections, remove no-op Atomics.notify, aspect guard,
         transformPoint w-divide guard
Phase 6: IMU calibration timeout, unsubscribe returns, bias correction consistency,
         LightingEstimator typeof check
Phase 7: dtolnay/rust-toolchain fix, track Cargo.lock, SDK build in CI, example
         import paths, opt-level=3
Phase 8: Multiplicative hash for vocabulary, removed duplicate constructor

543 Rust tests pass, 493/494 SDK tests pass (1 pre-existing jsdom issue).

+++ src/features/nms.rs
@@ -22,7 +22,7 @@ pub fn non_maximum_suppression(keypoints: &[KeyPoint], radius: u32) -> Vec<KeyPo
         return keypoints.to_vec();
     }
 
-    let radius_squared = radius * radius;
+    let radius_squared = (radius as u64) * (radius as u64);
 
     // Sort by score descending
     let mut sorted: Vec<KeyPoint> = keypoints.to_vec();
@@ -41,7 +41,7 @@ pub fn non_maximum_suppression(keypoints: &[KeyPoint], radius: u32) -> Vec<KeyPo
 
         // Suppress all weaker keypoints within radius
         for j in (i + 1)..sorted.len() {
-            if !suppressed[j] && kp.distance_squared(&sorted[j]) <= radius_squared {
+            if !suppressed[j] && (kp.distance_squared(&sorted[j]) as u64) <= radius_squared {
                 suppressed[j] = true;
             }
         }
@@ -75,7 +75,7 @@ pub fn non_maximum_suppression_grid(
 
     let cols = width.div_ceil(cell_size);
     let rows = height.div_ceil(cell_size);
-    let grid_size = (cols * rows) as usize;
+    let grid_size = (cols as usize) * (rows as usize);
 
     // Track the best keypoint for each cell
     let mut best_per_cell: Vec<Option<KeyPoint>> = vec![None; grid_size];
